pyproject.toml
[tool.poetry]
name = "photo_sorter"
version = "0.1.0"
description = "Python application to sort photos based on meta data."
authors = ["Brun0-42"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-sugar = "*"
pytest-cov = "*"
mypy = "*"
isort = "*"
tox = "^4.11.4"
loguru = "*"
piexif = "*"
pillow = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
